feat: add copy link.
